Complete Organizational Overhaul -  Modules Appropriately Installed - Browserify, Watchify, and Exorcist Installed - Bundle.js Built - Fully Functional
