+++ index.html
@@ -208,6 +208,6 @@
   </div>
   </main>
   <script src="js/data.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/bundle/bundle.js"></script>
 </body>
 </html>
