+++ js/bundle/bundle.map.js
@@ -1,0 +1,39 @@
+{
+  "version": 3,
+  "sources": [
+    "../../../../../usr/lib/node_modules/browserify/node_modules/browser-pack/_prelude.js",
+    "main.js",
+    "modules/delete-report.js",
+    "modules/handle-view-navigation.js",
+    "modules/new-data-view-FL.js",
+    "modules/new-data-view-render.js",
+    "modules/new-main-html-list.js",
+    "modules/reports-page-render-FL.js",
+    "modules/reports-page-render.js",
+    "modules/show-delete-modal.js",
+    "modules/src-update.js",
+    "modules/submit-function.js",
+    "modules/switch-view.js",
+    "modules/user-search.js"
+  ],
+  "names": [],
+  "mappings": "AAAA;ACAA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACvEA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;AC3BA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACfA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;AC3BA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACjCA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACvFA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACjBA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACtBA;AACA;AACA;AACA;AACA;AACA;AACA;;ACNA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACTA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;AC/EA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;;ACXA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA;AACA",
+  "file": "generated.js",
+  "sourceRoot": "",
+  "sourcesContent": [
+    "(function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c=\"function\"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error(\"Cannot find module '\"+i+\"'\");throw a.code=\"MODULE_NOT_FOUND\",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u=\"function\"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()",
+    "\n// Modules\nvar srcUpdate = require('./modules/src-update');\nvar deleteReport = require('./modules/delete-report');\nvar handleViewNavigation = require('./modules/handle-view-navigation');\nvar DOMDataViewForLoop = require('./modules/new-data-view-FL');\nvar reportsPageRenderForLoop = require('./modules/reports-page-render-FL');\nvar mainHTMLDOMNewListSubmissionRender = require('./modules/new-main-html-list');\nvar showDeleteModal = require('./modules/show-delete-modal');\nvar submitFunction = require('./modules/submit-function');\nvar userSearch = require('./modules/user-search');\n\n// Event Targets\nvar $searchForm = document.querySelector('.search-form');\nvar $form = document.querySelector('.entry-form-submit');\nvar $photoUrl = document.querySelector('.photo-url');\nvar $firstCityUl = document.querySelector('.first-city-ul');\nvar $secondCityUl = document.querySelector('.second-city-ul');\nvar mainElement = document.querySelector('.main');\nvar $searchNavbar = document.querySelector('.search-anchor');\nvar $createNewReportButton = document.querySelector('.create-new-report-button');\nvar $goBackButton = document.querySelector('.go-back-button-event');\nvar $reportsNavbar = document.querySelector('.reports-anchor');\nvar $reportsCityNameSelectorAll = document.querySelectorAll('.reports-city-name');\nvar $deleteReportYes = document.querySelector('.delete-report-yes');\n\n// Event Listeners\n$photoUrl.addEventListener('input', srcUpdate);\n$form.addEventListener('submit', submitFunction);\n$searchForm.addEventListener('submit', userSearch);\n$deleteReportYes.addEventListener('click', deleteReport);\n$reportsNavbar.addEventListener('click', handleViewNavigation);\n$searchNavbar.addEventListener('click', handleViewNavigation);\n$createNewReportButton.addEventListener('click', handleViewNavigation);\n$goBackButton.addEventListener('click', handleViewNavigation);\n$deleteReportYes.addEventListener('click', deleteReport);\n$reportsNavbar.addEventListener('click', handleViewNavigation);\n$searchNavbar.addEventListener('click', handleViewNavigation);\n$createNewReportButton.addEventListener('click', handleViewNavigation);\n$goBackButton.addEventListener('click', handleViewNavigation);\n\n// Loops\nfor (var a = 0; a < data.cities.length; a++) {\n  var city = data.cities[a];\n  mainElement.append(DOMDataViewForLoop(city));\n  if (data.cities.length === 1) {\n    $firstCityUl.append(reportsPageRenderForLoop(data.cities[a]));\n  } else if (a % 2 === 0) {\n    $secondCityUl.append(reportsPageRenderForLoop(data.cities[a]));\n  } else {\n    $firstCityUl.append(reportsPageRenderForLoop(data.cities[a]));\n  }\n}\n\nfor (var w = 0; w < data.entries.length; w++) {\n  const $ulSelectorAll = document.querySelectorAll('ul');\n  for (var x = 0; x < $ulSelectorAll.length; x++) {\n    if ($ulSelectorAll[x].getAttribute('class') === data.entries[w].cityName) {\n      $ulSelectorAll[x].append(mainHTMLDOMNewListSubmissionRender(data.entries[w]));\n    }\n  }\n}\n\nfor (var b = 0; b < $reportsCityNameSelectorAll.length; b++) {\n  $reportsCityNameSelectorAll[b].addEventListener('click', handleViewNavigation);\n}\n\nvar $deleteReportTextSelectorAll = document.querySelectorAll('.delete-report-text');\nfor (var n = 0; n < $deleteReportTextSelectorAll.length; n++) {\n  $deleteReportTextSelectorAll[n].addEventListener('click', showDeleteModal);\n}\n",
+    "\nvar switchView = require('./switch-view');\nconst $deleteModal = document.querySelector('.cancel-background');\n\nfunction deleteReport(event) {\n  for (var i = 0; i < data.entries.length; i++) {\n    var $liSelectorAll = document.querySelectorAll('li');\n    var nodeListArray = [];\n    for (var a = 0; a < $liSelectorAll.length; a++) {\n      if (data.entries[i].entryId === parseInt($liSelectorAll[a].getAttribute('data-entry-id'))) {\n        var $cityUlsSelectorAll = document.querySelectorAll('ul');\n        for (var b = 0; b < $cityUlsSelectorAll.length; b++) {\n          if ($liSelectorAll[a].getAttribute('city-id') === $cityUlsSelectorAll[b].getAttribute('class')) {\n            nodeListArray.push($liSelectorAll[a]);\n            $cityUlsSelectorAll[b].removeChild($liSelectorAll[a]);\n            data.entries.splice(i, 1);\n            $deleteModal.className = 'cancel-background flex justify-center position-fixed hidden';\n            switchView(event.target.getAttribute('data-view'));\n          }\n\n        }\n      }\n    }\n  }\n}\n\nmodule.exports = deleteReport;\n",
+    "var switchView = require('./switch-view');\n\nvar $form = document.querySelector('.entry-form-submit');\nvar $formPlaceholderImg = document.querySelector('.placeholder-img');\nvar $searchForm = document.querySelector('.search-form');\n\nfunction handleViewNavigation(event) {\n  $formPlaceholderImg.setAttribute('src', 'images/placeholder-image-square.jpg');\n  $form.reset();\n  $searchForm.reset();\n  var buttonDataView = event.target.getAttribute('data-view');\n  switchView(buttonDataView);\n}\n\nmodule.exports = handleViewNavigation;\n",
+    "function DOMDataViewForLoop(cityName) {\n  var dataViewDiv = document.createElement('div');\n  var containerDiv = document.createElement('div');\n  var rowDiv = document.createElement('div');\n  var columnFullDiv = document.createElement('div');\n  var h1ListHeading = document.createElement('h1');\n  var cityNameUl = document.createElement('ul');\n  var h1ListHeadingTextContent = document.createTextNode(cityName);\n\n  dataViewDiv.setAttribute('data-view', cityName);\n  dataViewDiv.setAttribute('class', 'view hidden');\n  containerDiv.setAttribute('class', 'container');\n  rowDiv.setAttribute('class', 'row');\n  columnFullDiv.setAttribute('class', 'column-full text-align-center');\n  h1ListHeading.setAttribute('class', 'list-heading');\n  h1ListHeading.appendChild(h1ListHeadingTextContent);\n  cityNameUl.setAttribute('class', cityName);\n\n  columnFullDiv.append(h1ListHeading, cityNameUl);\n  rowDiv.append(columnFullDiv);\n  containerDiv.append(rowDiv);\n  dataViewDiv.append(containerDiv);\n\n  return dataViewDiv;\n}\n\nmodule.exports = DOMDataViewForLoop;\n",
+    "var handleViewNavigation = require('./handle-view-navigation');\n\nvar $cityResultName = document.querySelector('.city-name');\n\nfunction mainHTMLDOMNewDataViewSubmissionRender() {\n  var dataViewDiv = document.createElement('div');\n  var containerDiv = document.createElement('div');\n  var rowDiv = document.createElement('div');\n  var columnFullDiv = document.createElement('div');\n  var h1ListHeading = document.createElement('h1');\n  var cityNameUl = document.createElement('ul');\n  var h1ListHeadingTextContent = document.createTextNode($cityResultName.textContent);\n\n  dataViewDiv.setAttribute('data-view', $cityResultName.textContent);\n  dataViewDiv.setAttribute('class', 'view hidden');\n  containerDiv.setAttribute('class', 'container');\n  rowDiv.setAttribute('class', 'row');\n  columnFullDiv.setAttribute('class', 'column-full text-align-center');\n  h1ListHeading.setAttribute('class', 'list-heading');\n  h1ListHeading.appendChild(h1ListHeadingTextContent);\n  cityNameUl.setAttribute('class', $cityResultName.textContent);\n\n  columnFullDiv.append(h1ListHeading, cityNameUl);\n  rowDiv.append(columnFullDiv);\n  containerDiv.append(rowDiv);\n  dataViewDiv.append(containerDiv);\n\n  dataViewDiv.addEventListener('click', handleViewNavigation);\n\n  return dataViewDiv;\n}\n\nmodule.exports = mainHTMLDOMNewDataViewSubmissionRender;\n",
+    "function mainHTMLDOMNewListSubmissionRender(entry) {\n\n  var mainRowLi = document.createElement('li');\n  var firstColumnHalfDiv = document.createElement('div');\n  var secondColumnHalfDiv = document.createElement('div');\n  var submittedImg = document.createElement('img');\n  var timeDateTitleDiv = document.createElement('div');\n  var listTitleDiv = document.createElement('div');\n  var titleH2 = document.createElement('h2');\n  var listTimeDateDiv = document.createElement('div');\n  var dateH3 = document.createElement('h3');\n  var timeH3 = document.createElement('h3');\n  var descriptionRowDiv = document.createElement('div');\n  var listDescriptionP = document.createElement('p');\n  var wasWeatherRowDiv = document.createElement('div');\n  var wasWeatherDiv = document.createElement('div');\n  var wasWeatherH4 = document.createElement('h4');\n  var wasWeatherPDiv = document.createElement('div');\n  var wasWeatherP = document.createElement('p');\n  var titleH2TextContent = document.createTextNode(entry.photoTitleValue);\n  var dateH3TextContent = document.createTextNode(entry.date);\n  var timeh3TextContent = document.createTextNode(entry.time);\n  var listDescriptionPTextContent = document.createTextNode(entry.commentsValue);\n  var wasWeatherH4TextContent = document.createTextNode('Was the official weather report accurate for this day?');\n  var wasWeatherPTextContent = document.createTextNode(entry.radioChecked);\n  var deleteReportRowDiv = document.createElement('div');\n  var firstColumnHalfDeleteRow = document.createElement('div');\n  var secondColumnHalfDeleteRow = document.createElement('div');\n  var deleteRowTextA = document.createElement('a');\n  var deleteRowTextContent = document.createTextNode('Delete This Report');\n\n  mainRowLi.setAttribute('class', 'row flex-wrap-wrapped');\n  mainRowLi.setAttribute('data-entry-id', entry.entryId);\n  mainRowLi.setAttribute('city-id', entry.cityName);\n  firstColumnHalfDiv.setAttribute('class', 'column-half');\n  secondColumnHalfDiv.setAttribute('class', 'column-half');\n  submittedImg.setAttribute('src', entry.photoUrlValue);\n  submittedImg.setAttribute('alt', 'user-submitted-img');\n  submittedImg.setAttribute('class', 'list-img');\n  timeDateTitleDiv.setAttribute('class', 'list-time-date-and-title row align-items-center');\n  listTitleDiv.setAttribute('class', 'list-title column-75');\n  titleH2.setAttribute('class', 'margin-block-unset');\n  titleH2.appendChild(titleH2TextContent);\n  listTimeDateDiv.setAttribute('class', 'list-time-and-date column-25');\n  dateH3.setAttribute('class', 'margin-block-unset');\n  dateH3.appendChild(dateH3TextContent);\n  timeH3.setAttribute('class', 'margin-block-unset');\n  timeH3.appendChild(timeh3TextContent);\n  descriptionRowDiv.setAttribute('class', 'row');\n  listDescriptionP.setAttribute('class', 'list-description column-full margin-block-unset');\n  listDescriptionP.appendChild(listDescriptionPTextContent);\n  wasWeatherRowDiv.setAttribute('class', 'was-weather row');\n  wasWeatherDiv.setAttribute('class', 'column-75 flex align-items-center');\n  wasWeatherH4.setAttribute('class', 'was-weather-h4');\n  wasWeatherH4.appendChild(wasWeatherH4TextContent);\n  wasWeatherPDiv.setAttribute('class', 'column-25 flex align-items-center justify-content-center');\n  wasWeatherP.setAttribute('class', 'was-weather-p');\n  wasWeatherP.appendChild(wasWeatherPTextContent);\n  deleteReportRowDiv.setAttribute('class', 'delete-report row border-bottom-white width-100p');\n  firstColumnHalfDeleteRow.setAttribute('class', 'column-half');\n  secondColumnHalfDeleteRow.setAttribute('class', 'column-half text-align-right');\n  deleteRowTextA.setAttribute('class', 'delete-report-text margin-block-unset');\n  deleteRowTextA.appendChild(deleteRowTextContent);\n\n  firstColumnHalfDiv.append(submittedImg);\n\n  listTitleDiv.append(titleH2);\n  listTimeDateDiv.append(timeH3, dateH3);\n  timeDateTitleDiv.append(listTitleDiv, listTimeDateDiv);\n\n  descriptionRowDiv.append(listDescriptionP);\n\n  wasWeatherDiv.append(wasWeatherH4);\n  wasWeatherPDiv.append(wasWeatherP);\n  wasWeatherRowDiv.append(wasWeatherDiv, wasWeatherPDiv);\n\n  secondColumnHalfDeleteRow.append(deleteRowTextA);\n  deleteReportRowDiv.append(firstColumnHalfDeleteRow, secondColumnHalfDeleteRow);\n\n  secondColumnHalfDiv.append(timeDateTitleDiv, descriptionRowDiv, wasWeatherRowDiv, deleteReportRowDiv);\n\n  mainRowLi.append(firstColumnHalfDiv, secondColumnHalfDiv);\n\n  return mainRowLi;\n}\n\nmodule.exports = mainHTMLDOMNewListSubmissionRender;\n",
+    "function reportsPageRenderForLoop(cityName) {\n  var li = document.createElement('li');\n  var h2 = document.createElement('h2');\n\n  li.setAttribute('data-view', cityName);\n  li.setAttribute('class', 'reports-city-name-holder');\n  h2.setAttribute('class', 'reports-city-name');\n  h2.setAttribute('data-view', cityName);\n  var h2TextContent = cityName;\n\n  h2.append(h2TextContent);\n  li.append(h2);\n\n  return li;\n}\n\nmodule.exports = reportsPageRenderForLoop;\n",
+    "var handleViewNavigation = require('./handle-view-navigation');\nvar $cityResultName = document.querySelector('.city-name');\n\nfunction reportsPageRender() {\n  var li = document.createElement('li');\n  var h2 = document.createElement('h2');\n  var h2TextContent = document.createTextNode($cityResultName.textContent);\n\n  li.setAttribute('data-view', $cityResultName.textContent);\n  li.setAttribute('class', 'reports-city-name-holder');\n  h2.setAttribute('class', 'reports-city-name');\n  h2.setAttribute('data-view', $cityResultName.textContent);\n  h2.appendChild(h2TextContent);\n  li.append(h2);\n\n  li.addEventListener('click', handleViewNavigation);\n  h2.addEventListener('click', handleViewNavigation);\n\n  return li;\n}\n\nmodule.exports = reportsPageRender;\n",
+    "const $deleteModal = document.querySelector('.cancel-background');\nfunction showDeleteModal() {\n  $deleteModal.className = 'cancel-background flex justify-content-center position-fixed view';\n}\n\nmodule.exports = showDeleteModal;\n",
+    "var $formPlaceholderImg = document.querySelector('.placeholder-img');\nvar $photoUrl = document.querySelector('.photo-url');\n\nfunction srcUpdate(event) {\n  $formPlaceholderImg.setAttribute('src', $photoUrl.value);\n\n}\n\nmodule.exports = srcUpdate;\n",
+    "\nvar reportsPageRender = require('./reports-page-render');\nvar mainHTMLDOMNewListSubmissionRender = require('./new-main-html-list');\nvar mainHTMLDOMNewDataViewSubmissionRender = require('./new-data-view-render');\nvar switchView = require('./switch-view');\n\nvar $cityResultName = document.querySelector('.city-name');\nvar $form = document.querySelector('.entry-form-submit');\nvar $cityResultTime = document.querySelector('.city-time');\nconst XHRDate = '';\nvar $firstCityUl = document.querySelector('.first-city-ul');\nvar mainElement = document.querySelector('.main');\nvar $secondCityUl = document.querySelector('.second-city-ul');\n\nfunction submitFunction(event) {\n  event.preventDefault();\n  var cityName = $cityResultName.textContent;\n  var photoTitleValue = $form.elements.photoTitle.value;\n  var commentsValue = $form.elements.comments.value;\n  var photoUrlValue = $form.elements.photoUrl.value;\n  var radioChecked = $form.elements.choice.value;\n  var time = $cityResultTime.textContent;\n  var date = XHRDate;\n  var submissionObject = { cityName, photoTitleValue, photoUrlValue, commentsValue, radioChecked, time, date };\n\n  if (data.entries.length === 0) {\n    data.cities.push(submissionObject.cityName);\n    $firstCityUl.appendChild(reportsPageRender());\n    mainElement.appendChild(mainHTMLDOMNewDataViewSubmissionRender());\n    const $ulSelectorAll = document.querySelectorAll('ul');\n    for (var c = 0; c < $ulSelectorAll.length; c++) {\n      if ($ulSelectorAll[c].getAttribute('class') === submissionObject.cityName) {\n        $ulSelectorAll[c].append(mainHTMLDOMNewListSubmissionRender(submissionObject));\n      }\n    }\n    submissionObject.entryId = data.nextEntryId;\n    data.nextEntryId++;\n    data.entries.unshift(submissionObject);\n    $form.reset();\n    switchView('search-bar');\n    return;\n  }\n  for (var i = 0; i < data.cities.length; i++) {\n    if (data.cities[i] === submissionObject.cityName) {\n      submissionObject.entryId = data.nextEntryId;\n      const $ulSelectorAll = document.querySelectorAll('ul');\n      for (var d = 0; d < $ulSelectorAll.length; d++) {\n        if ($ulSelectorAll[d].getAttribute('class') === submissionObject.cityName) {\n          $ulSelectorAll[d].append(mainHTMLDOMNewListSubmissionRender(submissionObject));\n        }\n      }\n      data.nextEntryId++;\n      data.entries.unshift(submissionObject);\n      $form.reset();\n      switchView('search-bar');\n      return;\n    }\n  }\n  data.cities.push(submissionObject.cityName);\n  submissionObject.entryId = data.nextEntryId;\n  if (data.cities.length % 2 === 0) {\n    $secondCityUl.append(reportsPageRender());\n  } else {\n    $firstCityUl.append(reportsPageRender());\n  }\n  mainElement.appendChild(mainHTMLDOMNewDataViewSubmissionRender());\n  const $ulSelectorAll = document.querySelectorAll('ul');\n  for (var f = 0; f < $ulSelectorAll.length; f++) {\n    if ($ulSelectorAll[f].getAttribute('class') === submissionObject.cityName) {\n      $ulSelectorAll[f].append(mainHTMLDOMNewListSubmissionRender(submissionObject));\n    }\n  }\n  data.nextEntryId++;\n  data.entries.unshift(submissionObject);\n  $form.reset();\n  switchView('search-bar');\n}\n\nmodule.exports = submitFunction;\n",
+    "function switchView(viewName) {\n  var $viewSelectorAll = document.querySelectorAll('.view');\n  for (var i = 0; i < $viewSelectorAll.length; i++) {\n    if ($viewSelectorAll[i].getAttribute('data-view') === viewName) {\n      $viewSelectorAll[i].className = 'view';\n    } else {\n      $viewSelectorAll[i].className = 'view hidden';\n    }\n  }\n}\nmodule.exports = switchView;\n",
+    "var switchView = require('./switch-view');\n\nvar $cityResultName = document.querySelector('.city-name');\nvar $cityResultTime = document.querySelector('.city-time');\nvar $weatherStatusResult = document.querySelector('.weather-status');\nvar $fahrenheitResult = document.querySelector('.fahrenheit-td');\nvar $humidityResult = document.querySelector('.humidity-td');\nvar $gustResult = document.querySelector('.gust-td');\nvar $cloudCoverageResult = document.querySelector('.cloud-cover-td');\nvar $feelsLikeResult = document.querySelector('.feels-like-td');\nvar $visibilityResult = document.querySelector('.visibility-td');\nvar $windMphResult = document.querySelector('.wind-mph-td');\nvar $precipitationResult = document.querySelector('.precipitation-td');\nvar $searchForm = document.querySelector('.search-form');\nvar $searchInput = document.querySelector('.search-input');\n// eslint-disable-next-line no-unused-vars\nlet XHRDate = '';\n\nfunction userSearch(event) {\n  event.preventDefault();\n  var xhr = new XMLHttpRequest();\n  xhr.open('GET', 'https://api.weatherapi.com/v1/current.json?key=182e266b7561494e81d230926211611&q=' + $searchInput.value + '&aqi=no');\n  xhr.responseType = 'json';\n  xhr.addEventListener('load', function () {\n    var XHRLocalTime = xhr.response.location.localtime;\n    var timeAndDateString = '';\n    timeAndDateString += '' + XHRLocalTime + '';\n    var timeAndDateArray = [];\n    var currentTimeOrDate = '';\n\n    for (var a = 0; a < timeAndDateString.length; a++) {\n      if (timeAndDateString[a] === ' ') {\n        timeAndDateArray.push(currentTimeOrDate);\n        currentTimeOrDate = '';\n      } else {\n        currentTimeOrDate += timeAndDateString[a];\n      }\n\n    }\n    timeAndDateArray.push(currentTimeOrDate);\n\n    var weatherStatus = Object.values(xhr.response.current.condition)[0];\n\n    $cityResultName.textContent = xhr.response.location.name + ', ' + xhr.response.location.region;\n    $cityResultTime.textContent = timeAndDateArray[1] + ' UTC';\n    XHRDate = timeAndDateArray[0];\n    $weatherStatusResult.textContent = weatherStatus;\n    $fahrenheitResult.textContent = xhr.response.current.temp_f + '\\xB0' + ' ' + 'Fahrenheit';\n    $feelsLikeResult.textContent = 'Feels like ' + xhr.response.current.feelslike_f + '\\xB0' + ' Fahreinheit';\n    $gustResult.textContent = xhr.response.current.gust_mph + 'mph' + ' Gust';\n    $cloudCoverageResult.textContent = xhr.response.current.cloud + '%' + ' Cloud Coverage';\n    $visibilityResult.textContent = xhr.response.current.vis_miles + ' Miles Visibility';\n    $windMphResult.textContent = xhr.response.current.wind_mph + 'mph' + ' Wind';\n    $precipitationResult.textContent = xhr.response.current.precip_in + 'in' + ' Precipitation';\n    $humidityResult.textContent = xhr.response.current.humidity + '%' + ' Humidity';\n  });\n  xhr.send();\n  $searchForm.reset();\n  switchView('search-bar-result');\n\n}\n\nmodule.exports = userSearch;\n"
+  ]
+}
