+++ js/data.js
@@ -1,4 +1,3 @@
-/* exported data */
 var data = {
   view: 'submissions',
   entries: [],
