+++ js/modules/handle-view-navigation.js
@@ -1,0 +1,15 @@
+var switchView = require('./switch-view');
+
+var $form = document.querySelector('.entry-form-submit');
+var $formPlaceholderImg = document.querySelector('.placeholder-img');
+var $searchForm = document.querySelector('.search-form');
+
+function handleViewNavigation(event) {
+  $formPlaceholderImg.setAttribute('src', 'images/placeholder-image-square.jpg');
+  $form.reset();
+  $searchForm.reset();
+  var buttonDataView = event.target.getAttribute('data-view');
+  switchView(buttonDataView);
+}
+
+module.exports = handleViewNavigation;
