+++ js/modules/new-data-view-FL.js
@@ -1,0 +1,27 @@
+function DOMDataViewForLoop(cityName) {
+  var dataViewDiv = document.createElement('div');
+  var containerDiv = document.createElement('div');
+  var rowDiv = document.createElement('div');
+  var columnFullDiv = document.createElement('div');
+  var h1ListHeading = document.createElement('h1');
+  var cityNameUl = document.createElement('ul');
+  var h1ListHeadingTextContent = document.createTextNode(cityName);
+
+  dataViewDiv.setAttribute('data-view', cityName);
+  dataViewDiv.setAttribute('class', 'view hidden');
+  containerDiv.setAttribute('class', 'container');
+  rowDiv.setAttribute('class', 'row');
+  columnFullDiv.setAttribute('class', 'column-full text-align-center');
+  h1ListHeading.setAttribute('class', 'list-heading');
+  h1ListHeading.appendChild(h1ListHeadingTextContent);
+  cityNameUl.setAttribute('class', cityName);
+
+  columnFullDiv.append(h1ListHeading, cityNameUl);
+  rowDiv.append(columnFullDiv);
+  containerDiv.append(rowDiv);
+  dataViewDiv.append(containerDiv);
+
+  return dataViewDiv;
+}
+
+module.exports = DOMDataViewForLoop;
