+++ js/modules/new-data-view-render.js
@@ -1,0 +1,33 @@
+var handleViewNavigation = require('./handle-view-navigation');
+
+var $cityResultName = document.querySelector('.city-name');
+
+function mainHTMLDOMNewDataViewSubmissionRender() {
+  var dataViewDiv = document.createElement('div');
+  var containerDiv = document.createElement('div');
+  var rowDiv = document.createElement('div');
+  var columnFullDiv = document.createElement('div');
+  var h1ListHeading = document.createElement('h1');
+  var cityNameUl = document.createElement('ul');
+  var h1ListHeadingTextContent = document.createTextNode($cityResultName.textContent);
+
+  dataViewDiv.setAttribute('data-view', $cityResultName.textContent);
+  dataViewDiv.setAttribute('class', 'view hidden');
+  containerDiv.setAttribute('class', 'container');
+  rowDiv.setAttribute('class', 'row');
+  columnFullDiv.setAttribute('class', 'column-full text-align-center');
+  h1ListHeading.setAttribute('class', 'list-heading');
+  h1ListHeading.appendChild(h1ListHeadingTextContent);
+  cityNameUl.setAttribute('class', $cityResultName.textContent);
+
+  columnFullDiv.append(h1ListHeading, cityNameUl);
+  rowDiv.append(columnFullDiv);
+  containerDiv.append(rowDiv);
+  dataViewDiv.append(containerDiv);
+
+  dataViewDiv.addEventListener('click', handleViewNavigation);
+
+  return dataViewDiv;
+}
+
+module.exports = mainHTMLDOMNewDataViewSubmissionRender;
