+++ js/modules/reports-page-render-FL.js
@@ -1,0 +1,17 @@
+function reportsPageRenderForLoop(cityName) {
+  var li = document.createElement('li');
+  var h2 = document.createElement('h2');
+
+  li.setAttribute('data-view', cityName);
+  li.setAttribute('class', 'reports-city-name-holder');
+  h2.setAttribute('class', 'reports-city-name');
+  h2.setAttribute('data-view', cityName);
+  var h2TextContent = cityName;
+
+  h2.append(h2TextContent);
+  li.append(h2);
+
+  return li;
+}
+
+module.exports = reportsPageRenderForLoop;
