+++ js/modules/src-update.js
@@ -1,0 +1,9 @@
+var $formPlaceholderImg = document.querySelector('.placeholder-img');
+var $photoUrl = document.querySelector('.photo-url');
+
+function srcUpdate(event) {
+  $formPlaceholderImg.setAttribute('src', $photoUrl.value);
+
+}
+
+module.exports = srcUpdate;
