+++ js/modules/switch-view.js
@@ -1,0 +1,11 @@
+function switchView(viewName) {
+  var $viewSelectorAll = document.querySelectorAll('.view');
+  for (var i = 0; i < $viewSelectorAll.length; i++) {
+    if ($viewSelectorAll[i].getAttribute('data-view') === viewName) {
+      $viewSelectorAll[i].className = 'view';
+    } else {
+      $viewSelectorAll[i].className = 'view hidden';
+    }
+  }
+}
+module.exports = switchView;
