+++ package.json
@@ -37,7 +37,6 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
-
   },
   "license": "ISC",
   "dependencies": {}
